Update sdk unread messages demo
